arm cortex vector_table: use initializer list to fill the vector table; remove attributes from static template variable
Attributes `used` and `section(".isr_vector")` are not hard coded in
vector_table any more, these have to be explicitely set in a
compilation unit. While this makes it more versatile, the main reason
for this change is that some compilers do not honor these attributes
within templates.

With gcc 4.x, the section attribute was honored within (implicit
instantiaded) static template member variables. To create the vector
table in section ".isr_vector" with the old code, the following
(simple) statement was used:

    mptl::vector_table< &_stack_top, resources > vector_table;

As of gcc 5.x (as well as clang), this does not work any more, section
attributes are not honored for static template variables. The new way
of creating arm vector tables in section ".isr_vector" would now be:

    const auto isr_vector __attribute__((used, section(".isr_vector")))
      = mptl::vector_table< &_stack_top, resources >::value;

This might look more complicated, but it makes it more obvious why
vector table is in section ".isr_vector" when reading the code.